Conflicts:
	app/src/main/java/com/shootr/android/util/ErrorMessageFactory.java
	app/src/main/res/values/strings.xml

See merge request !65

+++ domain/src/main/java/com/shootr/android/domain/interactor/stream/RemoveFromFavoritesInteractor.java
@@ -40,10 +40,11 @@ public class RemoveFromFavoritesInteractor implements Interactor {
 
     @Override public void execute() throws Exception {
         Favorite existingFavorite = localFavoriteRepository.getFavoriteByStream(idStream);
-        checkNotNull(existingFavorite);
-        localFavoriteRepository.removeFavoriteByStream(existingFavorite.getIdStream());
-        notifyCompleted();
-        remoteFavoriteRepository.removeFavoriteByStream(existingFavorite.getIdStream());
+        if (existingFavorite != null) {
+            localFavoriteRepository.removeFavoriteByStream(existingFavorite.getIdStream());
+            notifyCompleted();
+            remoteFavoriteRepository.removeFavoriteByStream(existingFavorite.getIdStream());
+        }
     }
 
     protected void notifyCompleted() {
